REFACTOR: Arquivo db_utils.py movido de app/core para app/utils
